feat(consensus): implement HFC era history state machine and slot-to-time conversion
Replace hardcoded era lookup tables with a proper Hard Fork Combinator (HFC)
era history state machine in torsten-consensus. This introduces:

- EraHistory type tracking era boundaries with exact slot/epoch/time arithmetic,
  following Haskell's Ouroboros.Consensus.HardFork.History.Summary model
- Genesis-driven initialization from Byron + Shelley genesis parameters
- Dynamic era transition detection during sync (block.era changes propagated
  to EraHistory via pending_era_transition on LedgerState)
- Slot-to-wallclock and wallclock-to-slot conversions across all era boundaries
- Slot-to-epoch and epoch-to-first-slot lookups across era boundaries
- Era::from_protocol_major() for governance-driven era transitions
- u128 time_pico to handle mainnet Byron times exceeding u64::MAX
- CBOR bignum encoding (tag 2) for large picosecond values

Replaces:
- shelley_transition_epoch_for_magic() hardcoded lookup (kept for bootstrap)
- current_wall_clock_slot() manual Byron+Shelley arithmetic (~80 lines)
- build_era_summaries() 200+ line hardcoded mainnet/testnet era tables

Net: -316 lines removed, +262 lines added (+ 1058 line era_history module
with 14 comprehensive unit tests).

Closes #303

+++ crates/torsten-node/src/node/n2c_query/types.rs
@@ -237,8 +237,9 @@ pub struct EraSummary {
     pub start_slot: u64,
     /// Start epoch number
     pub start_epoch: u64,
-    /// Start time in picoseconds relative to system start
-    pub start_time_pico: u64,
+    /// Start time in picoseconds relative to system start.
+    /// Uses u128 because mainnet Byron end time (~9e19 pico) exceeds u64::MAX.
+    pub start_time_pico: u128,
     /// End bound (None = current/unbounded era)
     pub end: Option<EraBound>,
     /// Epoch length in slots
@@ -256,7 +257,8 @@ pub struct EraSummary {
 pub struct EraBound {
     pub slot: u64,
     pub epoch: u64,
-    pub time_pico: u64,
+    /// Picoseconds from system start (u128 to avoid overflow for mainnet Byron).
+    pub time_pico: u128,
 }
 
 /// Stake snapshot result (mark/set/go)
